lambda/requirements.txt
Pillow==9.5.0
boto3
